backend/requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
python-multipart
PyPDF2
pydantic
google-generativeai